Update EdgeControlBase.cs to support new ReversePath property
Added support for reversing the path geometry.  This support allows the user to specify a property value to indicate that the path geometry should as specified should be reversed.  This change is helpful when animating a shape along the path and have a requirement to reverse the direction of the animation (without using AutoReverse property of Storyboard for example)